draws somethig without affines

+++ 3DViewer_2/parser.h
@@ -1,23 +1,19 @@
 #ifndef SRC_PARSER_H_
 #define SRC_PARSER_H_
 
-
-#include <iostream>
 #include <fstream>
+#include <iostream>
+#include <memory>
 #include <sstream>
-
 #include <string>
 #include <vector>
-#include <memory>
-
-#include "parser.h"
 
 struct Model {
-  Model(): vertices{}, faces{}, scaler{} {}
-  Model(Model&& other) noexcept :
-    vertices(std::move(other.vertices)),
-    faces(std::move(other.faces)),
-    scaler(other.scaler) {}
+  Model() : vertices{}, faces{}, scaler{} {}
+  Model(Model&& other) noexcept
+      : vertices(std::move(other.vertices)),
+        faces(std::move(other.faces)),
+        scaler(other.scaler) {}
 
   Model(const Model& other) = delete;
 
@@ -32,19 +28,20 @@ struct Model {
 
   Model& operator=(const Model& other) = delete;
 
-  std::vector<double> vertices;
-  std::vector<int> faces;
+  std::vector<float> vertices;
+  std::vector<unsigned int> faces;
   double scaler;
 };
 
 class ObjParser {
  public:
   Model Parse(const std::string& filename);
+
  private:
-  Model ParseData(const std::string& filename); 
+  Model ParseData(const std::string& filename);
   Model CalculateScaler(Model& model);
 };
 
 std::ostream& operator<<(std::ostream& os, const Model& model);
 
-#endif // SRC_PARSER_H_
+#endif  // SRC_PARSER_H_
